Add files via upload
Claude_tkr5.py is the latest version of a tkinter interface for doing HEX OPERATIONS , with reversing of text allowed, file creation and database creation and insertion of all data. Filenames might not be a problem, one more revision coming soon to make sure the names are all logical and unique and are based on the input filenames. Folder creation and organization by function this way is still required to understand how the files were created .