Add files via upload
this script can browse directly from a file dialog for two input files, A and B , and is capable of reversing the order of the two selected input files, reversing the text of either or both input files, and saves the result file as a file named in a logical way. The order of the base filenames is the same as the order of the concatenated files data and the suffix "rev' is added to the end of an input files base filename if the text data is reversed. Thus the data needed to understand the processing of the input file is preserved in the output filename.